NEW: refactor of bridge tests
