+++ systemtest/src/test/java/io/strimzi/systemtest/bridge/HttpBridgeST.java
@@ -1,0 +1,108 @@
+/*
+ * Copyright 2019, Strimzi authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+package io.strimzi.systemtest.bridge;
+
+import io.strimzi.api.kafka.model.KafkaResources;
+import io.strimzi.systemtest.Constants;
+import io.strimzi.systemtest.HttpBridgeBaseST;
+import io.vertx.core.json.JsonArray;
+import io.vertx.core.json.JsonObject;
+import io.vertx.junit5.VertxExtension;
+import org.apache.kafka.clients.consumer.ConsumerConfig;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+
+import static io.strimzi.systemtest.Constants.BRIDGE;
+import static io.strimzi.systemtest.Constants.REGRESSION;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+@Tag(BRIDGE)
+@Tag(REGRESSION)
+@ExtendWith(VertxExtension.class)
+class HttpBridgeST extends HttpBridgeBaseST {
+    private static final Logger LOGGER = LogManager.getLogger(HttpBridgeST.class);
+
+    private String bridgeHost = "";
+    private int bridgePort = Constants.HTTP_BRIDGE_DEFAULT_PORT;
+
+    @Test
+    void testSendSimpleMessage() throws Exception {
+        int messageCount = 50;
+        String topicName = "topic-simple-send";
+        // Create topic
+        testClassResources.topic(CLUSTER_NAME, topicName).done();
+
+        JsonObject records = generateHttpMessages(messageCount);
+        JsonObject response = sendHttpRequests(records, bridgeHost, bridgePort, topicName);
+        checkSendResponse(response, messageCount);
+        receiveMessagesExternal(NAMESPACE, topicName, messageCount);
+    }
+
+    @Test
+    void testReceiveSimpleMessage() throws Exception {
+        int messageCount = 50;
+        String topicName = "topic-simple-receive";
+        // Create topic
+        testClassResources.topic(CLUSTER_NAME, topicName).done();
+
+        String name = "my-kafka-consumer";
+        String groupId = "my-group";
+
+        JsonObject config = new JsonObject();
+        config.put("name", name);
+        config.put("format", "json");
+        config.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
+        // Create consumer
+        JsonObject response = createBridgeConsumer(config, bridgeHost, bridgePort, groupId);
+        assertThat("Consumer wasn't created correctly", response.getString("instance_id"), is(name));
+        // Create topics json
+        JsonArray topic = new JsonArray();
+        topic.add(topicName);
+        JsonObject topics = new JsonObject();
+        topics.put("topics", topic);
+        // Subscribe
+        assertTrue(subscribeHttpConsumer(topics, bridgeHost, bridgePort, groupId, name));
+        // Send messages to Kafka
+        sendMessagesExternal(NAMESPACE, topicName, messageCount);
+        // Try to consume messages
+        JsonArray bridgeResponse = receiveHttpRequests(bridgeHost, bridgePort, groupId, name);
+        if (bridgeResponse.size() == 0) {
+            // Real consuming
+            bridgeResponse = receiveHttpRequests(bridgeHost, bridgePort, groupId, name);
+        }
+        assertThat("Sent message count is not equal with received message count", bridgeResponse.size(), is(messageCount));
+        // Delete consumer
+        assertTrue(deleteConsumer(bridgeHost, bridgePort, groupId, name));
+    }
+
+    @BeforeAll
+    void createClassResources() {
+        LOGGER.info("Deploy Kafka and Kafka Bridge before tests");
+        // Deploy kafka
+        testClassResources.kafkaEphemeral(CLUSTER_NAME, 1, 1)
+                .editSpec()
+                .editKafka()
+                .editListeners()
+                .withNewKafkaListenerExternalNodePort()
+                .withTls(false)
+                .endKafkaListenerExternalNodePort()
+                .endListeners()
+                .endKafka()
+                .endSpec().done();
+
+        // Deploy http bridge
+        testClassResources.kafkaBridge(CLUSTER_NAME, KafkaResources.plainBootstrapAddress(CLUSTER_NAME), 1, Constants.HTTP_BRIDGE_DEFAULT_PORT).done();
+
+        deployBridgeNodePortService();
+        bridgePort = getBridgeNodePort();
+        bridgeHost = kubeClient(NAMESPACE).getNodeAddress();
+    }
+}
