+++ systemtest/src/test/java/io/strimzi/systemtest/bridge/HttpBridgeTlsST.java
@@ -2,11 +2,15 @@
  * Copyright 2019, Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.systemtest;
+package io.strimzi.systemtest.bridge;
 
-import io.fabric8.kubernetes.api.model.Service;
+import io.strimzi.api.kafka.model.CertSecretSource;
 import io.strimzi.api.kafka.model.KafkaResources;
-import io.strimzi.systemtest.utils.StUtils;
+import io.strimzi.api.kafka.model.KafkaUser;
+import io.strimzi.api.kafka.model.listener.KafkaListenerAuthenticationTls;
+import io.strimzi.api.kafka.model.listener.KafkaListenerTls;
+import io.strimzi.systemtest.Constants;
+import io.strimzi.systemtest.HttpBridgeBaseST;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.junit5.VertxExtension;
@@ -18,29 +22,24 @@ import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-import java.util.HashMap;
-import java.util.Map;
-
 import static io.strimzi.systemtest.Constants.BRIDGE;
 import static io.strimzi.systemtest.Constants.REGRESSION;
-import static io.strimzi.systemtest.Resources.getSystemtestsServiceResource;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @Tag(BRIDGE)
 @Tag(REGRESSION)
 @ExtendWith(VertxExtension.class)
-public class HttpBridgeST extends HttpBridgeBaseST {
-    private static final Logger LOGGER = LogManager.getLogger(HttpBridgeST.class);
-    public static final String NAMESPACE = "bridge-cluster-test";
+class HttpBridgeTlsST extends HttpBridgeBaseST {
+    private static final Logger LOGGER = LogManager.getLogger(HttpBridgeTlsST.class);
 
     private String bridgeHost = "";
     private int bridgePort = Constants.HTTP_BRIDGE_DEFAULT_PORT;
+    private String userName = "bob";
 
     @Test
-    void testSendSimpleMessage() throws Exception {
+    void testSendSimpleMessageTls() throws Exception {
         int messageCount = 50;
         String topicName = "topic-simple-send";
         // Create topic
@@ -48,20 +47,12 @@ public class HttpBridgeST extends HttpBridgeBaseST {
 
         JsonObject records = generateHttpMessages(messageCount);
         JsonObject response = sendHttpRequests(records, bridgeHost, bridgePort, topicName);
-        JsonArray offsets = response.getJsonArray("offsets");
-        assertEquals(messageCount, offsets.size());
-        for (int i = 0; i < messageCount; i++) {
-            JsonObject metadata = offsets.getJsonObject(i);
-            assertEquals(0, metadata.getInteger("partition"));
-            assertEquals(i, metadata.getLong("offset"));
-            LOGGER.debug("offset size: {}, partition: {}, offset size: {}", offsets.size(), metadata.getInteger("partition"), metadata.getLong("offset"));
-        }
-
-        receiveMessagesConsole(NAMESPACE, topicName, messageCount);
+        checkSendResponse(response, messageCount);
+        receiveMessagesExternalTls(NAMESPACE, topicName, messageCount, userName);
     }
 
     @Test
-    void testReceiveSimpleMessage() throws Exception {
+    void testReceiveSimpleMessageTls() throws Exception {
         int messageCount = 50;
         String topicName = "topic-simple-receive";
         // Create topic
@@ -85,7 +76,7 @@ public class HttpBridgeST extends HttpBridgeBaseST {
         // Subscribe
         assertTrue(subscribeHttpConsumer(topics, bridgeHost, bridgePort, groupId, name));
         // Send messages to Kafka
-        sendMessagesConsole(NAMESPACE, topicName, messageCount);
+        sendMessagesExternalTls(NAMESPACE, topicName, messageCount, userName);
         // Try to consume messages
         JsonArray bridgeResponse = receiveHttpRequests(bridgeHost, bridgePort, groupId, name);
         if (bridgeResponse.size() == 0) {
@@ -99,47 +90,47 @@ public class HttpBridgeST extends HttpBridgeBaseST {
 
     @BeforeAll
     void createClassResources() {
-        LOGGER.info("Creating resources before the test class");
-        prepareEnvForOperator(NAMESPACE);
+        LOGGER.info("Deploy Kafka and Kafka Bridge before tests");
 
-        createTestClassResources();
-        applyRoleBindings(NAMESPACE);
-        // 050-Deployment
-        testClassResources.clusterOperator(NAMESPACE).done();
+        KafkaListenerAuthenticationTls auth = new KafkaListenerAuthenticationTls();
+        KafkaListenerTls listenerTls = new KafkaListenerTls();
+        listenerTls.setAuth(auth);
+
         // Deploy kafka
         testClassResources.kafkaEphemeral(CLUSTER_NAME, 1, 1)
                 .editSpec()
                 .editKafka()
                 .editListeners()
                 .withNewKafkaListenerExternalNodePort()
-                .withTls(false)
+                .withTls(true)
                 .endKafkaListenerExternalNodePort()
+                .withTls(listenerTls)
+                .withNewTls()
+                .endTls()
                 .endListeners()
                 .endKafka()
                 .endSpec().done();
 
-        Map<String, String> map = new HashMap<>();
-        map.put("strimzi.io/cluster", CLUSTER_NAME);
-        map.put("strimzi.io/kind", "KafkaBridge");
-        map.put("strimzi.io/name", CLUSTER_NAME + "-bridge");
+        // Create Kafka user
+        KafkaUser userSource = testClassResources.tlsUser(CLUSTER_NAME, userName).done();
+        waitTillSecretExists(userName);
+
+        // Initialize CertSecretSource with certificate and secret names for consumer
+        CertSecretSource certSecret = new CertSecretSource();
+        certSecret.setCertificate("ca.crt");
+        certSecret.setSecretName(clusterCaCertSecretName(CLUSTER_NAME));
 
         // Deploy http bridge
-        testClassResources.kafkaBridge(CLUSTER_NAME, KafkaResources.plainBootstrapAddress(CLUSTER_NAME), 1, Constants.HTTP_BRIDGE_DEFAULT_PORT).done();
-        // Create load balancer service for expose bridge outside openshift
-        Service service = getSystemtestsServiceResource(bridgeLoadBalancer, Constants.HTTP_BRIDGE_DEFAULT_PORT)
-                .editSpec()
-                .withType("NodePort")
-                .withSelector(map)
-                .endSpec().build();
-        testClassResources.createServiceResource(service, NAMESPACE).done();
-        StUtils.waitForNodePortService(bridgeLoadBalancer);
+        testClassResources.kafkaBridge(CLUSTER_NAME, KafkaResources.tlsBootstrapAddress(CLUSTER_NAME), 1, Constants.HTTP_BRIDGE_DEFAULT_PORT)
+            .editSpec()
+            .withNewTls()
+            .withTrustedCertificates(certSecret)
+            .endTls()
+            .endSpec().done();
 
-        Service extBootstrapService = kubeClient(NAMESPACE).getClient().services()
-                .inNamespace(NAMESPACE)
-                .withName(bridgeLoadBalancer)
-                .get();
+        deployBridgeNodePortService();
 
-        bridgePort = extBootstrapService.getSpec().getPorts().get(0).getNodePort();
-        bridgeHost = kubeClient(NAMESPACE).listNodes().get(0).getStatus().getAddresses().get(0).getAddress();
+        bridgePort = getBridgeNodePort();
+        bridgeHost = kubeClient(NAMESPACE).getNodeAddress();
     }
 }
